Revert "move ActionResult to models file"
This reverts commit f34774658f198c53f4eceed76b53fcd1ead7bc10.
