+++ great_expectations_cloud/agent/actions/__init__.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 from great_expectations_cloud.agent.actions.agent_action import (
+    ActionResult,
     AgentAction,
     CreatedResource,
 )
