+++ great_expectations_cloud/agent/actions/data_assistants/utils.py
@@ -7,8 +7,10 @@ from typing import TYPE_CHECKING
 from great_expectations.datasource.fluent import Datasource as FluentDatasource
 from great_expectations.exceptions import DataContextError
 
-from great_expectations_cloud.agent.models import (
+from great_expectations_cloud.agent.actions.agent_action import (
     ActionResult,
+)
+from great_expectations_cloud.agent.models import (
     CreatedResource,
     RunDataAssistantEvent,
 )
