+++ great_expectations_cloud/agent/actions/draft_datasource_config_action.py
@@ -10,14 +10,14 @@ from great_expectations.datasource.fluent.interfaces import Datasource, TestConn
 from pydantic import v1 as pydantic_v1
 from typing_extensions import override
 
-from great_expectations_cloud.agent.actions import AgentAction
+from great_expectations_cloud.agent.actions import ActionResult, AgentAction
 from great_expectations_cloud.agent.config import (
     GxAgentEnvVars,
     generate_config_validation_error_text,
 )
 from great_expectations_cloud.agent.event_handler import register_event_action
 from great_expectations_cloud.agent.exceptions import ErrorCode, raise_with_error_code
-from great_expectations_cloud.agent.models import ActionResult, DraftDatasourceConfigEvent
+from great_expectations_cloud.agent.models import DraftDatasourceConfigEvent
 
 if TYPE_CHECKING:
     from great_expectations.compatibility.sqlalchemy.engine import Inspector
