+++ great_expectations_cloud/agent/actions/run_checkpoint.py
@@ -4,10 +4,12 @@ from typing import TYPE_CHECKING
 
 from typing_extensions import override
 
-from great_expectations_cloud.agent.actions.agent_action import AgentAction
+from great_expectations_cloud.agent.actions.agent_action import (
+    ActionResult,
+    AgentAction,
+)
 from great_expectations_cloud.agent.event_handler import register_event_action
 from great_expectations_cloud.agent.models import (
-    ActionResult,
     CreatedResource,
     RunCheckpointEvent,
     RunScheduledCheckpointEvent,
