+++ tests/agent/test_event_handler.py
@@ -12,6 +12,7 @@ from great_expectations.experimental.metric_repository.metrics import (
 from typing_extensions import override
 
 from great_expectations_cloud.agent.actions import (
+    ActionResult,
     AgentAction,
     DraftDatasourceConfigAction,
     ListTableNamesAction,
@@ -30,7 +31,6 @@ from great_expectations_cloud.agent.event_handler import (
     register_event_action,
 )
 from great_expectations_cloud.agent.models import (
-    ActionResult,
     DraftDatasourceConfigEvent,
     Event,
     EventBase,
